frontend/requirements.txt
click==7.1.2
streamlit
Pillow==8.0.1
requests_futures
python-multipart